Flip the default value of jit_compile=True for optimizer.
When TPUStrategy is used, we force jit_compile=False since TPU by default turns on XLA. Explicitly setting XLA on for optimizer causes unexpected behavior on TPU.

We also moved the check of variable key existence to base optimizer so that subclassed optimizers no longer need to copy the check around. Moreover, instead of raise KeyError, we log a warning and directly return because during model.save(), the update_step() will be retraced, and when XLA is on it will have key missing.

PiperOrigin-RevId: 436381960